% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activeGeomLayers.R
\name{activeGeomLayers}
\alias{activeGeomLayers}
\title{active geom layers}
\usage{
activeGeomLayers(ggplotObject)
}
\arguments{
\item{ggplotObject}{a ggplot object}
}
\description{
`activeGeomLayers` will return the geom layer index which can be active
}
\details{
`loon.ggplot` has an argument called `activeGeomLayers`. It is a vector to determine which geom layers can be active.
The default setting is `integer(0)`, however, `loon.ggplot` will automatically search the first `geom_histogram` or `geom_point` layer
to make it active. `activeGeomLayers` is more like a guide and give you an idea which one can be set as active.
}
\examples{
df <- data.frame(x = 1:3, y = 1:3, colour = c(1,3,5))
xgrid <- with(df, seq(min(x), max(x), length = 50))
interp <- data.frame(
  x = xgrid,
  y = approx(df$x, df$y, xout = xgrid)$y,
  colour = approx(df$x, df$colour, xout = xgrid)$y
)
p1 <- ggplot(data = df, aes(x, y, colour = colour)) +
  geom_line(interp, mapping = aes(x, y, colour = colour), size = 2) +
  geom_point(size = 5)
agL <- activeGeomLayers(p1)
loon.ggplot(p1, activeGeomLayers = agL$geom_point)


p2 <- ggplot(economics) +
  geom_rect(
    aes(xmin = start, xmax = end, fill = party),
    ymin = -Inf, ymax = Inf, alpha = 0.2,
    data = presidential
  ) +
  geom_text(
    aes(x = start, y = 2500, label = name),data = presidential,
    size = 3, vjust = 0, hjust = 0, nudge_x = 50
  ) +
  geom_line(aes(date, unemploy)) +
  scale_fill_manual(values = c("blue", "red"))
# none can be active
activeGeomLayers(p2)
#transparency is not allowed in tcltk
loon.ggplot(p2, ggGuides = TRUE, activeGeomLayers = integer(0))

}
\seealso{
\code{\link{loon.ggplot}}
}

