% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_serialaxes_layers.R
\name{add_serialaxes_layers}
\alias{add_serialaxes_layers}
\title{Layers on serial axes coordinate}
\usage{
add_serialaxes_layers(layer, plot, object)
}
\arguments{
\item{layer}{a layer object}

\item{plot}{a \code{ggplot} object}

\item{object}{some parameters used to modify this serial axes \code{ggplot} object (i.e. \code{axesLayout}, \code{axesLabels}, ...)}
}
\description{
Project the regular \code{geom} layers onto the serial axes coordinate.
}
