% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_getLocations-deprecated.R
\name{g_getLocations}
\alias{g_getLocations}
\alias{g_getLocations.default}
\alias{g_getLocations.l_pairs}
\title{get locations for \code{ggmatrix}}
\usage{
g_getLocations(target)

\method{g_getLocations}{default}(target)

\method{g_getLocations}{l_pairs}(target)
}
\arguments{
\item{target}{the (compound) loon plot whose locations are needed to lay out.}
}
\value{
a list of an appropriate subset of the named location arguments
`c("ncol", "nrow", "layout_matrix", "heights", "widths")`. \code{layout_matrix}
is an \code{nrow} by \code{ncol} matrix whose entries identify the location
of each plot in \code{g_getPlots()} by their index.
}
\description{
For the target compound \code{loon} plot, determines location in \code{ggmatrix}
}
\seealso{
\code{\link{l_getLocations}}, \code{\link{g_getPlots}}
}
