% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_getPlots.R
\name{g_getPlots}
\alias{g_getPlots}
\alias{g_getPlots.default}
\alias{g_getPlots.l_pairs}
\title{get \code{ggplot}s}
\usage{
g_getPlots(target)

\method{g_getPlots}{default}(target)

\method{g_getPlots}{l_pairs}(target)
}
\arguments{
\item{target}{the (compound) loon  plot to be laid out.}
}
\value{
a list of \code{ggplot}s.
}
\description{
For the target compound loon plot, determines all the \code{ggplot}s based on
the compound \code{loon} plot.
}
\seealso{
\code{\link{l_getPlots}}, \code{\link{g_getLocations}}
}
