% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{geom_imageGlyph-deprecated}
\alias{geom_imageGlyph-deprecated}
\alias{geom_imageGlyph}
\alias{geom_polygonGlyph}
\title{geom_imageGlyph (deprecated)}
\usage{
geom_imageGlyph(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  images,
  width = 4,
  height = 3,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_polygonGlyph(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  polygon_x,
  polygon_y,
  showArea = TRUE,
  linewidth = 1,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{images}{a list of images (a raster object, bitmap image). If not provided, a point visual (\code{geom_point()}) will be displayed.}

\item{width}{Numerical; width of image}

\item{height}{Numerical; height of image}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{polygon_x}{nested list of x-coordinates of polygons, one list element for each scatterplot point.
If not provided, a point visual (\code{geom_point()}) will be displayed.}

\item{polygon_y}{nested list of y-coordinates of polygons, one list element for each scatterplot point.
If not provided, a point visual (\code{geom_point()}) will be displayed.}

\item{showArea}{show area; deprecated now, please set `fill` or `colour` to control the shown area.}

\item{linewidth}{line width of the "glyph" object}
}
\description{
use \code{\link{geom_image_glyph}}, rather than \code{geom_imageGlyph}.

use \code{\link{geom_polygon_glyph}}, rather than \code{geom_polygonGlyph}.
}
\keyword{internal}
