% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-polygonGlyph.R
\name{geom_polygonGlyph}
\alias{geom_polygonGlyph}
\title{Add polygon glyph on scatter plot}
\usage{
geom_polygonGlyph(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., polygon_x, polygon_y, showArea = TRUE,
  linewidth = 1, na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to `layer()`.}

\item{polygon_x}{nested list of x-coordinates of polygons, one list element for each scatterplot point.
If not provided, `geom_point()` will be called.}

\item{polygon_y}{nested list of y-coordinates of polygons, one list element for each scatterplot point.
If not provided, `geom_point()` will be called.}

\item{showArea}{boolean, show a filled polygon or just the outline}

\item{linewidth}{linewidth of outline.}

\item{na.rm}{If FALSE, the default, missing values are removed with a warning.
If TRUE, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}
}
\description{
Add polygon glyph on scatter plot
}
\section{Aesthetics}{

geom_polygonGlyph() understands the following aesthetics (required aesthetics are in bold):
\itemize{
\item{\strong{x}}
\item{\strong{y}}
\item{alpha}
\item{colour}
\item{fill}
\item{group}
\item{shape}
\item{size}
\item{stroke}
\item{linetype}
}
}

\examples{
x_star <-
  c(-0.000864304235090734, 0.292999135695765, 0.949870354364736,
    0.474503025064823, 0.586862575626621, -0.000864304235090734,
    -0.586430423509075, -0.474070872947277, -0.949438202247191, -0.29256698357822)
y_star <-
  c(-1, -0.403630077787381, -0.308556611927398, 0.153846153846154,
    0.808556611927398, 0.499567847882455, 0.808556611927398,
    0.153846153846154, -0.308556611927398, -0.403630077787381)
x_cross <-
  c(-0.258931143762604, -0.258931143762604, -0.950374531835206,
    -0.950374531835206, -0.258931143762604, -0.258931143762604,
    0.259651397291847, 0.259651397291847, 0.948934024776722,
    0.948934024776722, 0.259651397291847, 0.259651397291847)
y_cross <-
  c(-0.950374531835206, -0.258931143762604, -0.258931143762604,
    0.259651397291847, 0.259651397291847, 0.948934024776722,
    0.948934024776722, 0.259651397291847, 0.259651397291847,
    -0.258931143762604, -0.258931143762604, -0.950374531835206)
x_hexagon <-
  c(0.773552290406223, 0, -0.773552290406223, -0.773552290406223,
    0, 0.773552290406223)
y_hexagon <-
  c(0.446917314894843, 0.894194756554307, 0.446917314894843,
    -0.447637568424085, -0.892754249495822, -0.447637568424085)
p <- ggplot(data = data.frame(x = 1:3, y = 1:3),
            mapping = aes(x = x, y = y)) +
  geom_polygonGlyph(polygon_x = list(x_star, x_cross, x_hexagon),
                    polygon_y = list(y_star, y_cross, y_hexagon),
                    colour = 'black', fill = 'red')
p
}
\seealso{
\code{\link{geom_serialAxesGlyph}}, \code{\link{geom_pointrangeGlyph}},
\code{\link{geom_imageGlyph}}, \code{\link{geom_textGlyph}}
}
