% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{geom_serialAxesGlyph-deprecated}
\alias{geom_serialAxesGlyph-deprecated}
\alias{geom_serialAxesGlyph}
\title{geom_serialAxesGlyph (deprecated)}
\usage{
geom_serialAxesGlyph(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  serialAxesData,
  sequence = NULL,
  linewidth = 1,
  scaling = c("variable", "data", "observation", "none"),
  axesLayout = c("parallel", "radial"),
  showAxes = FALSE,
  showArea = FALSE,
  showEnclosing = FALSE,
  axesColor = "black",
  bboxColor = "black",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{serialAxesData}{a serial axes numerical data set. If not provided, `geom_point()` will be called.}

\item{sequence}{A vector with variable names that defines the axes sequence}

\item{linewidth}{line width of the "glyph" object}

\item{scaling}{one of \code{data}, \code{variable}, \code{observation} or
\code{none} (not suggested the layout is the same with \code{data})
to specify how the data is scaled.}

\item{axesLayout}{either "radial" or "parallel"}

\item{showAxes}{boolean to indicate whether axes should be shown or not}

\item{showArea}{show area; deprecated now, please set `fill` or `colour` to control the shown area.}

\item{showEnclosing}{boolean to indicate whether enclosing should be shown or not}

\item{axesColor}{axes color}

\item{bboxColor}{bounding box color}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\description{
use \code{\link{geom_serialaxes_glyph}}, rather than \code{geom_serialAxesGlyph}.
}
\keyword{internal}
