% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unexportedFunctions.R
\name{get_scaledData}
\alias{get_scaledData}
\title{scale data}
\usage{
get_scaledData(
  data,
  sequence = NULL,
  scaling = c("variable", "data", "observation", "none"),
  displayOrder = NULL,
  as.data.frame = FALSE
)
}
\arguments{
\item{data}{A data frame}

\item{sequence}{vector with variable names that defines the axes sequence}

\item{scaling}{one of 'variable', 'data', 'observation' or 'none' to specify how the data is scaled.}

\item{displayOrder}{the order of the display}

\item{as.data.frame}{Return a matrix or a data.frame}
}
\description{
It is mainly used in serial axes
}
