% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_color_hue.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Default \code{ggplot} color palette}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{the number of colors}
}
\description{
Generate default discrete \code{ggplot} color
}
