% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_pipe.R
\name{gg_pipe}
\alias{gg_pipe}
\title{Create a ggplot object function for piping}
\usage{
gg_pipe(data, ggplotObject)
}
\arguments{
\item{data}{a data frame to use for ggplot}

\item{ggplotObject}{a ggplot object to be passed though}
}
\value{
a ggplot function, not used for plot, just for passing through
}
\description{
This function is used to pack and pass ggplot object through a pipe model (because of the
precedence of + and %>%, ggplot object will break the pipe model)
}
\examples{

library(magrittr)
g <- mtcars \%>\%
   gg_pipe(ggplot(aes(mpg, wt, colour = cyl)) + geom_point()) \%>\%
   loon.ggplot()
}
