% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_loon.R, R/ggplot2_loon_l_layer_graph.R,
%   R/ggplot2_loon_l_layer_histogram.R, R/ggplot2_loon_l_pairs.R,
%   R/ggplot2_loon_l_serialaxes.R, R/ggplot2_loon_l_ts.R
\name{ggplot2.loon}
\alias{ggplot2.loon}
\alias{ggplot2.loon.default}
\alias{ggplot2.loon.l_plot}
\alias{ggplot2.loon.l_hist}
\alias{ggplot2.loon.l_plot3D}
\alias{ggplot2.loon.l_layer_graph}
\alias{ggplot2.loon.l_layer_histogram}
\alias{ggplot2.loon.l_pairs}
\alias{ggplot2.loon.l_serialaxes}
\alias{ggplot2.loon.l_ts}
\title{Create a ggplot object from a loon object}
\usage{
ggplot2.loon(target, ...)

\method{ggplot2.loon}{default}(target, ...)

\method{ggplot2.loon}{l_plot}(target, ...)

\method{ggplot2.loon}{l_hist}(target, ...)

\method{ggplot2.loon}{l_plot3D}(target, ...)

\method{ggplot2.loon}{l_layer_graph}(target, ...)

\method{ggplot2.loon}{l_layer_histogram}(target, ...)

\method{ggplot2.loon}{l_pairs}(target, ...)

\method{ggplot2.loon}{l_serialaxes}(target, ...)

\method{ggplot2.loon}{l_ts}(target, ...)
}
\arguments{
\item{target}{either an object of class loon or a vector that specifies the
widget, layer, glyph, navigator or context completely.
The widget is specified by the widget path name (e.g. '.l0.plot'),
the remaining objects by their ids.}

\item{...}{arguments used in `ggplot2.loon()`, not for users}
}
\value{
a ggplot object
}
\description{
Get static ggplots from loon plots
}
\examples{
library(loon.ggplot)
l <- l_plot(iris, color = iris$Species)
p <- ggplot2.loon(l)
p # a ggplot object
str(p)
# add themes
p + ggtitle("iris") + theme_linedraw()

library(GGally)
p <- l_pairs(iris, showHistograms = TRUE)
g <- ggplot2.loon(p)
g + ggtitle("Iris pairs plot")
}
