% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggSerialaxes.R
\name{ggSerialAxes}
\alias{ggSerialAxes}
\title{ggplot with serialaxes}
\usage{
ggSerialAxes(ggObj, scaling = c("variable", "observation", "data",
  "none"), layout = c("parallel", "radial"), ...)
}
\arguments{
\item{ggObj}{A `ggplot` object}

\item{scaling}{one of 'variable', 'data', 'observation' or 'none' to
specify how the data is scaled. See Details for more information}

\item{layout}{either "radial" or "parallel"}

\item{...}{state arguments, see details}
}
\value{
a ggplot object
}
\description{
The ggplot serialaxes graphics displays multivariate data either as
a stacked star glyph plot, or as a parallel coordinate plot.
}
\details{
Available names set in `...`:
\itemize{
 \item{data: }{A data frame for serialaxes. If `NULL`, data must be set in `ggObj`}
 \item{axesLabels: }{A vector with variable names that defines the axes sequence.}
 \item{title: }{title of the display}
 \item{displayOrder: }{The display order of the observations}
 \item{color: }{Line color}
 \item{size: }{Line width}
 \item{showAxesLabels: }{Logical value to indicate whether axes labels should be shown or not}
 \item{showGuides: }{Logical value to indicate whether guides should be shown or not}
 \item{showAxes: }{Logical value to indicate whether axes should be shown or not}
 \item{showArea: }{Logical value to indicate whether to display lines or area}
}
}
\examples{
p <- ggplot(data = mtcars, mapping = aes(colour = as.factor(cyl))) \%>\%
       ggSerialAxes()
p
p +
  theme(legend.key = element_rect(fill = "lightblue", color = "black")) +
  scale_colour_manual(values = c("4" = "red", "6" = "blue", "8" = "green"))

}
