% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-Coord.R
\name{is.CoordPolar}
\alias{is.CoordPolar}
\title{Is polar coordinate system?}
\usage{
is.CoordPolar(coord)
}
\arguments{
\item{coord}{A \code{ggplot} object coordinate system}
}
\description{
Determine whether the \code{ggplot} object has polar coordinate system
}
