% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_ggplot.R
\name{is.l_ggplot}
\alias{is.l_ggplot}
\title{Reports whether x is a \code{l_ggplot} object}
\usage{
is.l_ggplot(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Reports whether x is a \code{l_ggplot} object
}
