% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{itemLabel-deprecated}
\alias{itemLabel-deprecated}
\alias{itemLabel}
\title{itemLabel (deprecated)}
\usage{
itemLabel(itemLabel = NULL, showItemLabels = NULL)
}
\arguments{
\item{itemLabel}{A character vector of length \code{n} with a string to be used to pop up when the
mouse hovers above that element.}

\item{showItemLabels}{A single logical value: \code{TRUE} if pop up labels are to appear on hover,
\code{FALSE} (the default) if they are not.}
}
\description{
use \code{\link{hover}}, rather than \code{itemLabel}.
}
\keyword{internal}
