% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loonGrob_l_ggplot.R
\name{l_get_arrangeGrobArgs.l_ggplot}
\alias{l_get_arrangeGrobArgs.l_ggplot}
\title{get arrangeGrob arguments}
\usage{
\method{l_get_arrangeGrobArgs}{l_ggplot}(target)
}
\arguments{
\item{target}{`l_ggplot` loon target.}
}
\value{
arrangeGrob arguments, see \code{\link{arrangeGrob}}
}
\description{
to get arrangeGrob arguments and used in \code{\link{loonGrob.l_compound}}
}
\examples{
p <- ggplot(mtcars, aes(mpg, wt, colour = cyl)) + geom_point()+ facet_wrap(~gear)
g <- loon.ggplot(p)
library(grid)
grid.newpage(); grid.draw(loonGrob(g))
}
\seealso{
\code{\link{loonGrob_layoutType.l_ggplot}}
}

