% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_getSubtitles.R
\name{l_getSubtitles}
\alias{l_getSubtitles}
\alias{l_getSubtitles.l_facet_ggplot}
\alias{l_getSubtitles.l_facet_wrap}
\alias{l_getSubtitles.l_facet_grid}
\title{Return the subtitles}
\usage{
l_getSubtitles(target)

\method{l_getSubtitles}{l_facet_ggplot}(target)

\method{l_getSubtitles}{l_facet_wrap}(target)

\method{l_getSubtitles}{l_facet_grid}(target)
}
\arguments{
\item{target}{an \code{l_facet_ggplot} object. If the ggplot object is
faceted (either by \code{facet_wrap} or \code{facet_grid}),
an \code{l_facet_ggplot} object will be returned once it is turned
to a \code{loon} plot.}
}
\value{
A list of labels, i.e. subtitles, labels, title, etc
}
\description{
Return the subtitles
}
\examples{
if(interactive()) {
p <- ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  facet_wrap(vars(class))
lp <- loon.ggplot(p)
l_getSubtitles(lp)
}

}
