% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_coords.R
\name{layout_coords}
\alias{layout_coords}
\title{layout matrix}
\usage{
layout_coords(target)
}
\arguments{
\item{target}{an object `ggplot2loon()` returns}
}
\description{
return the layout matrix of a list of loon plots
}
\examples{
p <- ggplot(mpg, aes(displ, cty)) + geom_point() + ggtitle("foo") +
 facet_grid(rows = vars(drv))
g <- ggplot2loon(p)
loon.ggplot:::layout_coords(g)
}
