% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_matrix.R
\name{layout_matrix}
\alias{layout_matrix}
\title{layout matrix}
\usage{
layout_matrix(target)
}
\arguments{
\item{target}{a list of loon object with specific names (x1y1, x1y2, ...)}
}
\description{
return the layout matrix of a list of loon plots
}
\examples{
p <- ggplot(mpg, aes(displ, cty)) + geom_point() + ggtitle("foo") +
 facet_grid(rows = vars(drv))
g <- loon.ggplot(p)
layout_matrix(g)
# l_getLocations(g)
}
