% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loon.ggplot-package.R
\docType{package}
\name{loon.ggplot-package}
\alias{loon.ggplot}
\alias{loon.ggplot-package}
\title{loon.ggplot: An add on package to loon that converts ggplot2 plots to interactive loon
    plots}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Convert ggplot object to interactive loon widgets, vice versa.
}
\author{
\strong{Maintainer}: Zehao Xu \email{z267xu@uwaterloo.ca}

Authors:
\itemize{
  \item R. Wayne Oldford \email{rwoldford@uwaterloo.ca}
}

}
\keyword{internal}
