% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loon.ggplot.R
\name{loon.ggplot}
\alias{loon.ggplot}
\title{Create a loon plot from a ggplot2 object}
\usage{
loon.ggplot(ggplotObject, ggGuides = FALSE, ...)
}
\arguments{
\item{ggplotObject}{a ggplot object}

\item{ggGuides}{if \code{TRUE}, loon plot will generate a ggplot guide layer. Default is FALSE.}

\item{...}{named arguments to modify loon plot states}
}
\value{
a loon widget
}
\description{
Interactive loon plots from ggplots
}
\examples{

 library(ggplot2)
 library(loon)
 p <- ggplot(mtcars, aes(mpg, wt, colour = cyl)) + geom_point()
 p
 g <- loon.ggplot(p)

 p <- ggplot(mpg, aes(class, hwy)) + geom_boxplot()
 g <- loon.ggplot(p, ggGuides = TRUE)
}

