% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loonGrob_l_ggplot.R
\name{loonGrob.l_ggplot}
\alias{loonGrob.l_ggplot}
\title{Create a grid grob from a loon widget handle}
\usage{
\method{loonGrob}{l_ggplot}(target, name = NULL, gp = NULL, vp = NULL)
}
\value{
a grid grob
}
\description{
Grid grobs are useful to create publication quality graphics.
}
\examples{
p <- ggplot(mtcars, aes(mpg, wt, colour = cyl)) + geom_point()+ facet_wrap(~gear)
g <- loon.ggplot(p)
library(grid)
grid.newpage(); grid.draw(loonGrob(g))
}

