% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loonGrob_l_ggplot.R
\name{loonGrob_layoutType.l_ggplot}
\alias{loonGrob_layoutType.l_ggplot}
\title{Create a grid grob from a loon widget handle}
\usage{
loonGrob_layoutType.l_ggplot(target)
}
\arguments{
\item{target}{loon target. \code{S3} method, see \link{loonGrob}}

\item{name}{a character identifier for the grob. Used to find the grob on the display list and/or as a child of another grob.}

\item{gp}{A gpar object, typically the output from a call to the function gpar. This is basically a list of graphical parameter settings.}

\item{vp}{a \link{viewport} object (or NULL).}
}
\value{
a grid grob
}
\description{
Grid grobs are useful to create publication quality graphics.
}
\examples{
p <- ggplot(mtcars, aes(mpg, wt, colour = cyl)) + geom_point()+ facet_wrap(~gear)
g <- loon.ggplot(p)
library(grid)
grid.newpage(); grid.draw(loonGrob(g))
}

