% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_loon_layers.R
\name{loonLayer}
\alias{loonLayer}
\title{Transform geom layers to loon layers}
\usage{
loonLayer(
  widget,
  layerGeom,
  data,
  ggplotPanelParams,
  ggObj,
  parent,
  label,
  ...
)
}
\arguments{
\item{widget}{a \code{loon} widget}

\item{layerGeom}{a \code{ggplot} \code{Geom} layer object}

\item{data}{a data frame (i.e. \code{x}, \code{y}, etc) of this particular layer}

\item{ggplotPanelParams}{\code{ggplot} panel parameters}

\item{ggObj}{the \code{ggplot} object}

\item{parent}{a valid Tk parent widget path.}

\item{label}{label used in the layers inspector}

\item{...}{not for users}
}
\description{
Function \code{loonLayer} is used to create \code{loon} non-interactive layers.
For some \code{ggplot2} extension packages, one can edit this function to realize the transformation.
}
