% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_chars.R
\docType{data}
\name{polygonGlyph}
\alias{polygonGlyph}
\alias{x_star}
\alias{y_star}
\alias{x_cross}
\alias{y_cross}
\alias{x_hexagon}
\alias{y_hexagon}
\alias{x_airplane}
\alias{y_airplane}
\title{Polygon glyph coordinates}
\format{
An object of class \code{character}.

An object of class \code{numeric} of length 10.

An object of class \code{numeric} of length 12.

An object of class \code{numeric} of length 12.

An object of class \code{numeric} of length 6.

An object of class \code{numeric} of length 6.

An object of class \code{numeric} of length 32.

An object of class \code{numeric} of length 32.
}
\usage{
x_star

y_star

x_cross

y_cross

x_hexagon

y_hexagon

x_airplane

y_airplane
}
\description{
Some useful polygon coordinates
}
\examples{
if(requireNamespace("grid")) {
  library(grid)
  grid.newpage()
  grid.polygon(x=(x_star + 1)/2,
               y=(1 - y_star)/2)
  grid.newpage()
  grid.polygon(x=(x_cross + 1)/2,
               y=(y_cross + 1)/2)
  grid.newpage()
  grid.polygon(x=(x_hexagon + 1)/2,
               y=(y_hexagon + 1)/2)
  grid.newpage()
  grid.polygon(x=(-x_airplane + 4)/10,
               y=(-y_airplane + 4)/10)
}
}
\seealso{
\link{geom_polygonGlyph}
}
\keyword{datasets}
