% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_ggplot.R
\name{print.l_ggplot}
\alias{print.l_ggplot}
\title{Explicitly draw plot}
\usage{
\method{print}{l_ggplot}(x, message = TRUE, ...)
}
\arguments{
\item{x}{plot to display}

\item{message}{logical; if \code{TRUE}, the way to create handle will be
printed out.}

\item{...}{other arguments used to modify function \code{ggplot2loon}}
}
\value{
Invisibly returns a \code{loon} widget
}
\description{
Explicitly draw plot
}
