% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_ggplot.R
\name{print.lggplot}
\alias{print.lggplot}
\title{Explicitly draw plot}
\usage{
\method{print}{lggplot}(x, ...)
}
\arguments{
\item{x}{plot to display}

\item{...}{other arguments used to modify function \code{ggplot2loon}}
}
\value{
Invisibly returns a \code{loon} widget
}
\description{
Explicitly draw plot
}
