% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_multi.R
\name{scaleBox}
\alias{scaleBox}
\title{Box scaling in 3D rotation}
\usage{
scaleBox(center = TRUE)
}
\arguments{
\item{center}{either a logical value or numeric-alike vector of length equal to the number of columns of
x, where 'numeric-alike' means that \code{as.numeric(.)} will be applied successfully
if \code{is.numeric(.)} is not true.}
}
\value{
A \code{trans} object
}
\description{
the variable is scaled to have equal ranges and,
when \code{center = TRUE}, to be centred by the average of the min and max.
}
\seealso{
\code{\link{l_scale3D}}
}
