% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive-ggproto.R
\name{scaleTo}
\alias{scaleTo}
\title{Change Plot Region}
\usage{
scaleTo(scaleTo = NULL, scaleToFun = NULL)
}
\arguments{
\item{scaleTo}{numerical; which layer to scale to}

\item{scaleToFun}{scale to function. See details.}
}
\value{
a \code{ggproto} object
}
\description{
Modify the \code{zoomX}, \code{zoomY}, \code{panX}, \code{panY},
etc to change the plot region
}
\details{
Argument \code{scaleTo} is used for additional plot region settings.
If the \code{scaleTo} is \code{NULL} (default), the region of the \code{loon}
is determined by the \code{ggplot} object
(i.e. \code{coord_cartesian}, \code{xlim}, etc);
else one can use \code{scaleToFun} to modify the region of the layer.

The \code{scaleToFun} is a function to scale the region.
If it is \code{NULL} (default), based on different layers, different scale functions
will be applied; else, users can select one that precisely tailor their own
problems. The table shows the available \code{scaleToFun} functions
\tabular{ll}{ \strong{scale to} \tab \strong{Subfunction}\cr
  plot \tab  \code{\link{l_scaleto_plot}}\cr
  world \tab \code{\link{l_scaleto_world}}\cr
  active \tab \code{\link{l_scaleto_active}}\cr
  selected \tab \code{\link{l_scaleto_selected}}\cr
  layer \tab \code{\link{l_scaleto_layer}}\cr}

Expect these, users can customize their own function. Note that, the arguments should match
the ones of functions shown in the table.
}
\examples{
if(interactive()) {
p <- l_ggplot(mtcars,
              mapping = aes(x = hp, y = mpg)) +
       geom_point(mapping = aes(color = factor(gear))) +
       geom_smooth(data = mtcars[mtcars$gear == 4, ],
                   method = "lm")
# a scatter plot with a fitted line on 4 gear cars
p
# scale to the second layer (smooth line)
p + scaleTo(scaleTo = 2)
# highlight the 3 gear cars
# scale to the selected points
p +
  selection(mtcars$gear == 3) +
  scaleTo(scaleTo = 1,
          scaleToFun = loon::l_scaleto_selected)
}
}
\seealso{
\code{\link{active}}, \code{\link{linking}}, \code{\link{selection}},
\code{\link{hover}}, \code{\link{interactivity}}
}
