% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{itemLabel-deprecated}
\alias{itemLabel-deprecated}
\alias{itemLabel}
\title{itemLabel (deprecated)}
\usage{
itemLabel(itemLabel = NULL, showItemLabels = NULL)
}
\arguments{
\item{itemLabel}{The customized querying information.}

\item{showItemLabels}{A logical value. Show item labels or not. Default is \code{FALSE}}
}
\description{
use \code{\link{hover}}, rather than \code{itemLabel}.
}
\keyword{internal}
