% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_coords.R
\name{layout_coords}
\alias{layout_coords}
\title{layout matrix}
\usage{
layout_coords(target)
}
\arguments{
\item{target}{an object \code{ggplot2loon()} returns}
}
\value{
a layout coordinate matrix
}
\description{
return the layout matrix of a list of loon plots
}
