% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{lggplot-deprecated}
\alias{lggplot-deprecated}
\alias{lggplot}
\title{Automatically create a loon widget (deprecated)}
\usage{
lggplot(data = NULL, mapping = aes(), ..., environment = parent.frame())
}
\arguments{
\item{data}{Default dataset to use for plot. If not already a data.frame, will be converted to one by fortify().
If not specified, must be supplied in each layer added to the plot.}

\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{...}{Other arguments passed on to methods. Not currently used.}

\item{environment}{DEPRECATED. Used prior to tidy evaluation.}
}
\description{
Automatically create a loon widget (deprecated)
}
\section{\code{lggplot}}{

For \code{lggplot}, use \code{\link{l_ggplot}}.
}

\keyword{internal}
