% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_ggplot.R
\name{lggplot}
\alias{lggplot}
\title{Automatically create a loon widget}
\usage{
lggplot(data = NULL, mapping = aes(), ..., environment = parent.frame())
}
\arguments{
\item{data}{Default dataset to use for plot. If not already a data.frame, will be converted to one by fortify().
If not specified, must be supplied in each layer added to the plot.}

\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{...}{Other arguments passed on to methods. Not currently used.}

\item{environment}{DEPRECATED. Used prior to tidy evaluation.}
}
\description{
It is retired. See \code{\link{l_ggplot}}
}
