% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loon.ggplot-package.R
\docType{package}
\name{loon.ggplot-package}
\alias{loon.ggplot-package}
\alias{_PACKAGE}
\title{loon.ggplot: An add on package to loon that converts ggplot2 plots to interactive loon
    plots}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

The \code{loon.ggplot} package brings both \code{loon} and \code{ggplot2} packages together. Data analysts who value the ease with which \code{ggplot2} can create meaningful graphics can now turn these \code{ggplot}s into interactive \code{loon} plots for more dynamic interaction with their data. Conversely, data analysts who explore data interactively can at any time turn a snapshot of their interactive \code{loon} plots into \code{ggplot}s. The former is accomplished by the simple translation function \code{ggplot2loon()} and the latter by the simple translation function \code{loon2ggplot()}.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/great-northern-diver/loon.ggplot/issues}
}

}
\author{
\strong{Maintainer}: Zehao Xu \email{z267xu@uwaterloo.ca}

Authors:
\itemize{
  \item R. Wayne Oldford \email{rwoldford@uwaterloo.ca}
}

}
\keyword{internal}
