% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loon.ggplot-package.R
\docType{package}
\name{loon.ggplot-package}
\alias{loon.ggplot-package}
\alias{_PACKAGE}
\title{loon.ggplot: Making 'ggplot2' Plots Interactive with 'loon' and Vice Versa}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

It provides a bridge between the 'loon' and  'ggplot2' packages. Data analysts who value the grammar pipeline provided by 'ggplot2' can turn these static plots into interactive 'loon' plots. Conversely, data analysts who explore data interactively with 'loon' can turn any 'loon' plot into a 'ggplot2' plot structure. The function 'loon.ggplot()' is applied to one plot structure will return the other.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/great-northern-diver/loon.ggplot/issues}
}

}
\author{
\strong{Maintainer}: Zehao Xu \email{z267xu@uwaterloo.ca}

Authors:
\itemize{
  \item R. Wayne Oldford \email{rwoldford@uwaterloo.ca}
}

}
\keyword{internal}
