% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loon.ggplot-package.R
\docType{package}
\name{loon.ggplot-package}
\alias{loon.ggplot-package}
\title{loon.ggplot: A Grammar of Interactive Graphics}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a bridge between the 'loon' and 'ggplot2' packages. Extends the grammar of ggplot to add clauses to create interactive 'loon' plots. Existing ggplot(s) can be turned into interactive 'loon' plots and 'loon' plots into static ggplot(s); the function 'loon.ggplot()' is the bridge from one plot structure to the other.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/great-northern-diver/loon.ggplot/issues}
}

}
\author{
\strong{Maintainer}: Zehao Xu \email{z267xu@gmail.com}

Authors:
\itemize{
  \item R. Wayne Oldford \email{rwoldford@uwaterloo.ca}
}

}
\keyword{internal}
