% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loon_ggplot.R
\name{loon.ggplot}
\alias{loon.ggplot}
\title{Create a loon plot from a ggplot2 object}
\usage{
loon.ggplot(ggplotObject, ggGuides = FALSE,
  activeGeomLayers = integer(0), tkLabels = FALSE, span = 5,
  canvasHeight = 850, canvasWidth = 700, ...)
}
\arguments{
\item{ggplotObject}{a ggplot object}

\item{ggGuides}{logical (default \code{FALSE}) to determine whether to draw a ggplot background or not.}

\item{activeGeomLayers}{to determine which geom layer is active. Only `geom_point()` and `geom_histogram()` can be set as active geom layer(s).
(Notice, more than one `geom_point()` layers can be set as active layers, but only one `geom_histogram()` can be set as an active geom layer)}

\item{tkLabels}{logical (default \code{TRUE}) to show tk packing labels or not}

\item{span}{the span of canvas}

\item{canvasHeight}{the height of canvas}

\item{canvasWidth}{the width of canvas}

\item{...}{named arguments to modify loon plot states}
}
\description{
Interactive loon plots from ggplots
}
\examples{
 p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
 g <- loon.ggplot(p)

 # show ggGuides
 p <- ggplot(mpg, aes(class, hwy)) + geom_boxplot()
 g <- loon.ggplot(p, ggGuides = TRUE)

 # tkLabels
p <- ggplot(mtcars) + geom_point(aes(x = wt, y = mpg,
   colour = factor(gear))) + facet_wrap(~am)
g1 <- loon.ggplot(p)
g2 <- loon.ggplot(p, tkLabels = TRUE)

\dontrun{
df <- data.frame(
  x = rnorm(120, c(0, 2, 4)),
  y = rnorm(120, c(1, 2, 1)),
  z = letters[1:3]
)
df2 <- dplyr::select(df, -z)
scatterplots <- ggplot(df, aes(x, y)) +
  geom_point(data = df2, colour = "grey70") +
  geom_point(aes(colour = z)) +
  facet_wrap(~z)
suppressWarnings(
  lp_scatterplots_active1 <- loon.ggplot(scatterplots, activeGeomLayers = 1)
)

lp_scatterplots_active2 <- loon.ggplot(scatterplots, activeGeomLayers = 2)
suppressWarnings(
 lp_scatterplots_active12 <- loon.ggplot(scatterplots, activeGeomLayers = c(1,2))
)
}
}
