% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loon_ggplot.R
\name{loon.ggplot}
\alias{loon.ggplot}
\title{Create a loon plot from a ggplot2 object}
\usage{
loon.ggplot(ggplotObject, ggGuides = FALSE, ...)
}
\arguments{
\item{ggplotObject}{a ggplot object}

\item{ggGuides}{if \code{TRUE}, loon plot will generate a ggplot guide layer. Default is FALSE.}

\item{...}{named arguments to modify loon plot states}
}
\value{
a loon widget
}
\description{
Interactive loon plots from ggplots
}
\examples{

 # set linkingGroup
 p1 <- ggplot(mtcars, aes(mpg, wt, colour = as.factor(cyl))) + geom_point()+ facet_wrap(~gear)
 p1
 g1 <- loon.ggplot(p1, linkingGroup = "mtcars")
 # loon histogram
 p2 <- ggplot(mtcars, aes(x = mpg, fill = as.factor(cyl))) + geom_histogram()
 p2
 g2 <- loon.ggplot(p2, linkingGroup = "mtcars")

 # show ggGuides
 p <- ggplot(mpg, aes(class, hwy)) + geom_boxplot()
 g <- loon.ggplot(p, ggGuides = TRUE)
}
