% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loon.ggplot.R
\name{loon.ggplot}
\alias{loon.ggplot}
\title{Create a loon plot from a ggplot2 object}
\usage{
loon.ggplot(ggplotObject, ...)
}
\arguments{
\item{ggplotObject}{a ggplot object}

\item{...}{named arguments to modify loon plot states}
}
\value{
a loon widget
}
\description{
Interactive loon plots from ggplots
}
\examples{

 library(ggplot2)
 library(loon)
 p <- ggplot(mtcars, aes(mpg, wt, colour = cyl)) + geom_point()
 p
 g <- loon.ggplot(p)

 df <- data.frame(x = 1:3, y = 1:3, colour = c(1,3,5))
 xgrid <- with(df, seq(min(x), max(x), length = 50))
 interp <- data.frame(
   x = xgrid,
   y = approx(df$x, df$y, xout = xgrid)$y,
   colour = approx(df$x, df$colour, xout = xgrid)$y)

 p <- ggplot(interp, aes(x, y, colour = colour)) +
   geom_line(size = 2) +
   geom_point(data = df, size = 5)
 p
 g <- loon.ggplot(p)
}

