% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loon_ggplot.R
\name{loon.ggplot}
\alias{loon.ggplot}
\title{Create a loon plot from a ggplot2 object}
\usage{
loon.ggplot(ggplotObject, activeGeomLayers = integer(0), parent = NULL,
  ggGuides = FALSE, pack = TRUE, tkLabels = NULL,
  exteriorLabelProportion = 1/5, canvasHeight = 700,
  canvasWidth = 850, ...)
}
\arguments{
\item{ggplotObject}{a ggplot object}

\item{activeGeomLayers}{to determine which geom layer is active. Only \code{geom_point()}
 and \code{geom_histogram()} can be set as active geom layer(s).
(Notice, more than one \code{geom_point()} layers can be set as active layers,
 but only one \code{geom_histogram()} can be set as an active geom layer)}

\item{parent}{parent widget path}

\item{ggGuides}{logical (default \code{FALSE}) to determine whether to draw a ggplot background or not.}

\item{pack}{logical (default \code{TRUE}) to pack widgets.
If \code{FALSE}, widgets will be produced but won't be packed and so will not appear in the display.}

\item{tkLabels}{logical (or \code{NULL}) to indicate whether the plot(s) are to be wrapped with
exterior labels (title, subtitle, xlabel or ylabel) using \code{tk.grid()}.
If \code{NULL} (default), then exterior labels appear only for multiple facets.
If \code{TRUE} exterior labels appear regardless; if \code{FALSE} no exterior labels appear.}

\item{exteriorLabelProportion}{space assigned to the vertical height/horizontal width of each exterior label
expressed as a proportion of a single plot's height/width.  Default is 0.2.
This is translated to a row/column span = 1 / exteriorLabelProportion for the plot size in
\code{tkgrid()}.}

\item{canvasHeight}{the height of canvas}

\item{canvasWidth}{the width of canvas}

\item{...}{named arguments to modify loon plot states}
}
\description{
Interactive loon plots from ggplots
}
\examples{
 p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
 g <- loon.ggplot(p)

 # show ggGuides
 p <- ggplot(mpg, aes(class, hwy)) + geom_boxplot()
 g <- loon.ggplot(p, ggGuides = TRUE)

 # tkLabels
p <- ggplot(mtcars) + geom_point(aes(x = wt, y = mpg,
   colour = factor(gear))) + facet_wrap(~am)
g1 <- loon.ggplot(p)
g2 <- loon.ggplot(p, tkLabels = FALSE)

\dontrun{
df <- data.frame(
  x = rnorm(120, c(0, 2, 4)),
  y = rnorm(120, c(1, 2, 1)),
  z = letters[1:3]
)
df2 <- dplyr::select(df, -z)
scatterplots <- ggplot(df, aes(x, y)) +
  geom_point(data = df2, colour = "grey70") +
  geom_point(aes(colour = z)) +
  facet_wrap(~z)
suppressWarnings(
  lp_scatterplots_active1 <- loon.ggplot(scatterplots, activeGeomLayers = 1)
)

lp_scatterplots_active2 <- loon.ggplot(scatterplots, activeGeomLayers = 2)

suppressWarnings(
 lp_scatterplots_active12 <- loon.ggplot(scatterplots, activeGeomLayers = c(1,2))
)
}
}
