% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loonGrob_l_ggplot.R
\name{loonGrob_layoutType.l_ggplot}
\alias{loonGrob_layoutType.l_ggplot}
\title{loonGrob layoutType}
\usage{
\method{loonGrob_layoutType}{l_ggplot}(target)
}
\arguments{
\item{target}{`l_ggplot` loon target.}
}
\value{
"arrangeGrobArgs"
}
\description{
a layout type, to be used in \code{\link{loonGrob.l_compound}}
}
\examples{
p <- ggplot(mtcars, aes(mpg, wt, colour = cyl)) + geom_point()+ facet_wrap(~gear)
g <- loon.ggplot(p)
library(grid)
grid.newpage(); grid.draw(loonGrob(g))
}
\seealso{
\code{\link{l_get_arrangeGrobArgs.l_ggplot}}
}

