% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{selecting-deprecated}
\alias{selecting-deprecated}
\alias{selecting}
\title{selecting (deprecated)}
\usage{
selecting(selected = NULL, selectBy = NULL, selectionLogic = NULL)
}
\arguments{
\item{selected}{A logical vector. If it is set as \code{TRUE}, the elements are highlighted
as the graphics are constructed. Default is \code{FALSE}}

\item{selectBy}{Select by "sweeping" (default) or "brushing".}

\item{selectionLogic}{Selection logic. One of "select" (default), "deselect" and "invert". See details.}
}
\description{
use \code{\link{selection}}, rather than \code{selecting}.
}
\keyword{internal}
